Update documentation to explicitly set Rail Env
If you don't specify the Rails environment when running the plugin
migration then it will use the default environment, which is often not
the one users want to use. Have had a couple of issues raised ( #33 and
#36 ) where people have found the migration failed because it was
running the migration on the Development rather than the Production
Rails environment.

Updated the documentation to say that you should explicitly set the
Rails Environment when running the plugin migration.